Syntax, duplicate dictionary lookups

+++ Rdmp.UI/ExtractionUIs/FilterUIs/ParameterUIs/ParameterCollectionUI.cs
@@ -252,20 +252,19 @@ public partial class ParameterCollectionUI : RDMPUserControl
 
         var deletables = olvParameters.SelectedObjects.OfType<IDeleteable>().ToArray();
 
-        if (!deletables.Any()) return;
-
-        var dr = Activator.YesNo(
-            deletables.Length == 1
-                ? $"Confirm deleting {deletables[0]}"
-                : $"Confirm deleting {deletables.Length} Parameters?", "Confirm delete");
-
-        if (!dr) return;
-
-        foreach (var d in deletables)
+        if (!deletables.Any() || e.KeyCode != Keys.Delete || !Activator.YesNo(
+                deletables.Length == 1
+                    ? $"Confirm deleting {deletables[0]}"
+                    : $"Confirm deleting {deletables.Length} Parameters?", "Confirm delete")) return;
+        foreach (IDeleteable d in olvParameters.SelectedObjects)
         {
             d.DeleteInDatabase();
             olvParameters.RemoveObject(d);
             Options.ParameterManager.RemoveParameter((ISqlParameter)d);
+
+            DisableRelevantObjects();
+            parameterEditorScintillaControl1.RegenerateSQL();
+            UpdateTabVisibility();
         }
 
         // Rebuild once after deleting all objects, instead of after each deletion
@@ -383,11 +382,11 @@ public partial class ParameterCollectionUI : RDMPUserControl
         e.StandardIcon = ToolTipControl.StandardIcons.Info;
 
         //if it is a problem parameter
-        if (parameterEditorScintillaControl1.ProblemObjects.TryGetValue(sqlParameter, out var o))
+        if (parameterEditorScintillaControl1.ProblemObjects.TryGetValue(sqlParameter, out var exception))
         {
             e.StandardIcon = ToolTipControl.StandardIcons.Warning;
             e.Title = "Problem Detected With Parameter";
-            e.Text = o.Message;
+            e.Text = exception.Message;
             return;
         }
